cleanup site meta tags
