+++ app/(pages)/not-found.tsx
@@ -12,6 +12,7 @@ import { LABELS } from "../labels"
 
 export const metadata: Metadata = {
   title: "404: Page Not Found",
+  description: "The requested page could not be found",
 }
 
 export default function NotFound() {
